+++ example/lib/main.dart
@@ -33,12 +33,20 @@ class _MyHomePageState extends State<MyHomePage> {
   final _items = <String, Todo>{};
   @override
   void initState() {
-    _db.collection('todos').stream.listen((event) {
+    _db.collection('todos').get().then((value) {
       setState(() {
-        final item = Todo.fromMap(event);
-        _items.putIfAbsent(item.id, () => item);
+        value?.entries.forEach((element) {
+          final item = Todo.fromMap(element.value);
+          _items.putIfAbsent(item.id, () => item);
+        });
       });
     });
+    // _db.collection('todos').stream.listen((event) {
+    //   setState(() {
+    //     final item = Todo.fromMap(event);
+    //     _items.putIfAbsent(item.id, () => item);
+    //   });
+    // });
     super.initState();
   }