+++ lib/src/collection_ref.dart
@@ -1,8 +1,7 @@
 part of localstore;
 
 /// A [CollectionRef] object can be used for adding documents, getting
-/// [DocumentRef]s, and querying for documents (using the methods
-/// inherited from [Query]).
+/// [DocumentRef]s, and querying for documents.
 class CollectionRef implements CollectionRefImpl {
   String _id;
 
@@ -14,23 +13,34 @@ class CollectionRef implements CollectionRefImpl {
 
   CollectionRef? _parent;
 
+  List<List>? _conditions;
+
   /// The parent [CollectionRef] of this document.
   CollectionRef? get parent => _parent;
 
-  CollectionRef._(this._id, [this._parent, this._delegate]);
+  CollectionRef._(this._id, [this._parent, this._delegate, this._conditions]);
   static final _cache = <String, CollectionRef>{};
 
   /// Returns an instance using the default [CollectionRef].
   factory CollectionRef(String id,
       [CollectionRef? parent, DocumentRef? delegate]) {
     final key = '${parent?.path ?? ''}${delegate?.id ?? ''}/$id/';
-    return _cache.putIfAbsent(key, () => CollectionRef._(id, parent, delegate));
+    final collectionRef =
+        _cache.putIfAbsent(key, () => CollectionRef._(id, parent, delegate));
+    collectionRef._conditions = null;
+    return collectionRef;
   }
 
   final _utils = Utils.instance;
+  final Map<String, dynamic?> _data = {};
 
   @override
-  Stream<Map<String, dynamic>> get stream => _utils.stream(path);
+  Stream<Map<String, dynamic>> get stream => _utils.stream(path, _conditions);
+
+  @override
+  Future<Map<String, dynamic>?> get() async {
+    return _data[path] ?? await _utils.get(path, true, _conditions);
+  }
 
   @override
   DocumentRef doc([String? id]) {
@@ -39,5 +49,25 @@ class CollectionRef implements CollectionRefImpl {
         .toRadixString(35)
         .substring(0, 9);
     return DocumentRef(id, this);
+  }
+
+  @override
+  CollectionRef where(
+    field, {
+    isEqualTo,
+  }) {
+    final conditions = <List>[];
+    void addCondition(dynamic field, String operator, dynamic value) {
+      List<dynamic> condition;
+
+      condition = <dynamic>[field, operator, value];
+      conditions.add(condition);
+    }
+
+    if (isEqualTo != null) addCondition(field, '==', isEqualTo);
+
+    _conditions = conditions;
+
+    return this;
   }
 }